backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
requests==2.31.0
newsapi-python==0.2.7